Undo refactoring
This reverts commit 42b4fe2ba20d228e17aa82b5faad121a0cfb4eab.
