+++ src/main/java/frc/team2412/robot/commands/shooter/ShooterTargetCommand.java
@@ -4,14 +4,14 @@ import static frc.team2412.robot.subsystem.ShooterSubsystem.ShooterConstants;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.team2412.robot.subsystem.ShooterSubsystem;
+import frc.team2412.robot.subsystem.ShooterVisionSubsystem;
 import frc.team2412.robot.util.ShooterDataDistancePoint;
-import frc.team2412.robot.util.ShooterVision;
 
 public class ShooterTargetCommand extends CommandBase {
     private final ShooterSubsystem shooter;
-    private final ShooterVision vision;
+    private final ShooterVisionSubsystem vision;
 
-    public ShooterTargetCommand(ShooterSubsystem shooter, ShooterVision vision) {
+    public ShooterTargetCommand(ShooterSubsystem shooter, ShooterVisionSubsystem vision) {
         this.shooter = shooter;
         this.vision = vision;
         addRequirements(shooter);
